Filename used when sending files is taken from hash key.
We want to send files generated as tempfiles with random names and
specify proper filename.

The key part of the filename hash was not used so we now use it as
the name of the file or files to upload.

This change is not backwards compatible if you were depending on
the name of the file sent.

BEFORE: Files were sent with its real name
    response = @session.post_multipart("/test", { :test_data => "123" }, { :test_file => "VERSION.yml" } )
AFTER: Files are sent with key as file name
    response = @session.post_multipart("/test", { :test_data => "123" }, { "VERSION.yml" => "VERSION.yml" } )
    response = @session.post_multipart("/test", { :test_data => "123" }, { "VERSION.yml" => "VERSION.yml", "GENERATED.txt" => "/tmp/icssuis501" } )